requirements.txt
Flask>=2.3
Flask-Cors>=3.0
openai>=1.14
ollama>=0.1.8
python-docx>=1.0
PyPDF2>=3.0
reportlab>=4.0
websocket-client>=1.6
python-pptx>=0.6.21

